+++ backend/services/models.py
@@ -126,6 +126,48 @@ class Service(models.Model, GeoItem):
             }
         }
 
+    @classmethod
+    def update_meili_documents(cls):
+        from search import client
+        published_services = cls.objects.filter(published=True)
+
+        settings = client.index('services').get_settings()
+
+        searchable_fields = settings.get('searchableAttributes', [])
+        if searchable_fields == ['*']:
+            raise NotImplemented
+
+        documents = []
+
+        for service in published_services:
+            flat_tags = [{
+                'id': tag.id,
+                tag.facet.translation_id: tag.value
+            } for tag in service.tags.all()]
+
+            fields = {
+                field: getattr(service, field)
+                for field in searchable_fields
+            }
+
+            documents.append({
+                'id': service.id,
+                **fields,
+                'tags': flat_tags
+            })
+
+        client.refresh('services', documents)
+
+    @classmethod
+    def update_meili_geodata(cls):
+        from search import client
+        subset_with_location = cls.objects.filter(
+            published=True, location__isnull=False)
+        
+        documents = [service.to_geodata() for service in subset_with_location]
+
+        client.refresh('geodata', documents, primary_key='slug')
+
     class Meta:
         ordering = ('id',)
